Create nshravankumar.Javascript
Start the React app using the npm start command and test the form by submitting with the email, name, and phone. The form should make a GET API call to check if the user exists or not. If the user exists, 
it should show an alert saying "User Found!". Otherwise, it should make a POST API call to create the user and show an alert saying "User Created Successfully!". If there's an error in any of the API calls,
it should show an alert saying "Error checking or creating user".